pyproject.toml
[tool.poetry]
name = "convert_markdown"
version = "0.1.0"
description = "Converte arquivos Markdown para DOCX formatado em ABNT."
authors = ["Brayan Pletsch <brayan.pletsch@gmail.com>"]
license = "MIT"
readme = { file = "README.md", content-type = "text/markdown" }
packages = [
  { include = "convert_markdown", from = "src" }
]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Text Processing :: Markup :: Markdown",
]
keywords = ["markdown", "docx", "abnt", "python-docx"]
include = ["LICENSE", "README.md"]

[tool.poetry.urls]
"Homepage"      = "https://github.com/BrayanPletsch/ConvertMarkdownToDocx"
"Repository"    = "https://github.com/BrayanPletsch/ConvertMarkdownToDocx"
"Documentation" = "https://github.com/BrayanPletsch/ConvertMarkdownToDocx#readme"
"Bug Tracker"   = "https://github.com/BrayanPletsch/ConvertMarkdownToDocx/issues"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
lxml = "5.3.2"
python-docx = "1.1.2"
typing_extensions = "4.13.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
flake8 = "^5.0"
isort = "^5.0"
black = "^23.0"

[tool.poetry.scripts]
convert_markdown = "convert_markdown.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"